requirements.txt
fastapi
uvicorn[standard]
httpx
sqlalchemy
pymysql
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
pydantic[email]
python-dotenv
